Added day 6
Pretty sure there's a much simpler way to do this with regexes, but couldn't figure it out faster enough so took the brute force approach. 

Worth noting in mind that Swift is odd and doesn't have a specific "give me the index this value appears at" function. Or it does, but Index is a type that's not a simple Int. So had to jump through some weird Range and Index hoops.